lct_back/requirements.txt
asgiref==3.7.2
Django==4.2.7
djangorestframework==3.14.0
psycopg2-binary==2.9.9
pytz==2023.3.post1
sqlparse==0.4.4
typing_extensions==4.8.0
rest_condition==1.0.3
pika==1.3.2
numpy==1.26.1
opencv-python==4.8.1.78
Pillow==10.1.0
python-dotenv==1.0.0